+++ dinky-client/dinky-client-base/src/main/java/org/dinky/parser/InsertSelectSqlParser.java
@@ -17,22 +17,28 @@
  *
  */
 
-package org.dinky.parse;
+package org.dinky.parser;
 
 /**
- * DeleteSqlParser
+ * InsertSelectSqlParser
  *
- * @since 2021/6/14 16:51
+ * @since 2021/6/14 16:53
  */
-public class DeleteSqlParser extends BaseSingleSqlParser {
+public class InsertSelectSqlParser extends BaseSingleSqlParser {
 
-    public DeleteSqlParser(String originalSql) {
+    public InsertSelectSqlParser(String originalSql) {
         super(originalSql);
     }
 
     @Override
     protected void initializeSegments() {
-        segments.add(new SqlSegment("(delete\\s+from)(.+)( where | ENDOFSQL)", "[,]"));
-        segments.add(new SqlSegment("(where)(.+)( ENDOFSQL)", "(and|or)"));
+        segments.add(new SqlSegment("(insert\\s+into)(.+)( select )", "[,]"));
+        segments.add(new SqlSegment("(select)(.+)(from)", "[,]"));
+        segments.add(new SqlSegment(
+                "(from)(.+?)( where | on | having | group\\s+by | order\\s+by | ENDOFSQL)",
+                "(,|\\s+left\\s+join\\s+|\\s+right\\s+join\\s+|\\s+inner\\s+join\\s+)"));
+        segments.add(new SqlSegment("(where|on|having)(.+?)( group\\s+by | order\\s+by | ENDOFSQL)", "(and|or)"));
+        segments.add(new SqlSegment("(group\\s+by)(.+?)( order\\s+by| ENDOFSQL)", "[,]"));
+        segments.add(new SqlSegment("(order\\s+by)(.+?)( ENDOFSQL)", "[,]"));
     }
 }
