+++ dinky-core/src/main/java/org/dinky/data/result/ResultBuilder.java
@@ -19,7 +19,7 @@
 
 package org.dinky.data.result;
 
-import org.dinky.parse.SqlType;
+import org.dinky.parser.SqlType;
 
 import org.apache.flink.table.api.TableResult;
