+++ dinky-executor/src/main/java/org/dinky/executor/Executor.java
@@ -26,7 +26,7 @@ import org.dinky.data.model.LineageRel;
 import org.dinky.data.result.SqlExplainResult;
 import org.dinky.interceptor.FlinkInterceptor;
 import org.dinky.interceptor.FlinkInterceptorResult;
-import org.dinky.parse.CustomParserImpl;
+import org.dinky.parser.CustomParserImpl;
 import org.dinky.utils.KerberosUtil;
 
 import org.apache.flink.api.common.ExecutionConfig;
