+++ dinky-executor/src/main/java/org/dinky/parse/SingleSqlParserFactory.java
@@ -19,6 +19,13 @@
 
 package org.dinky.parse;
 
+import org.dinky.parser.BaseSingleSqlParser;
+import org.dinky.parser.DeleteSqlParser;
+import org.dinky.parser.InsertSelectSqlParser;
+import org.dinky.parser.InsertSqlParser;
+import org.dinky.parser.SelectSqlParser;
+import org.dinky.parser.UpdateSqlParser;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
