+++ dinky-executor/src/test/java/org/dinky/parse/SqlTypeTest.java
@@ -19,6 +19,8 @@
 
 package org.dinky.parse;
 
+import org.dinky.parser.SqlType;
+
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
