+++ dinky-executor/src/test/java/org/dinky/trans/CreateTemporalTableFunctionParseStrategyTest.java
@@ -19,7 +19,7 @@
 
 package org.dinky.trans;
 
-import org.dinky.parse.CreateTemporalTableFunctionParseStrategy;
+import org.dinky.trans.parse.CreateTemporalTableFunctionParseStrategy;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
